Version 1.0
Finally! After 2 weeks since i made this repo, the newspaper is out (and i hope i get an A+ lol).

This release features 15 articles, a main page, a contact us page, a mission, vision and values page, embedded twitter posts, a script that creates random ads and also makes the cancel button on one of the pages work, lots of decoration and more!